Added an alternative initialization heuristic (minimal lookahead search).

+++ src/search/search_engine.h
@@ -115,11 +115,9 @@ protected:
     bool actionIsApplicable(ActionState const& action,
             State const& current) const {
         double res = 0.0;
-        for (unsigned int precondIndex = 0;
-             precondIndex < action.actionPreconditions.size();
-             ++precondIndex) {
-            action.actionPreconditions[precondIndex]->evaluate(res, current,
-                    action);
+        for (size_t precondIndex = 0; precondIndex < action.actionPreconditions.size(); ++precondIndex) {
+            action.actionPreconditions[precondIndex]->evaluate(
+                    res, current, action);
             if (MathUtils::doubleIsEqual(res, 0.0)) {
                 return false;
             }
@@ -322,16 +320,14 @@ protected:
         ActionHashMap::iterator it = applicableActionsCache.find(state);
         if (it != applicableActionsCache.end()) {
             assert(it->second.size() == res.size());
-            for (unsigned int i = 0; i < res.size(); ++i) {
+            for (size_t i = 0; i < res.size(); ++i) {
                 res[i] = it->second[i];
             }
         } else {
             if (hasUnreasonableActions) {
                 std::map<PDState, int, PDState::PDStateCompare> childStates;
 
-                for (unsigned int actionIndex = 0;
-                     actionIndex < numberOfActions;
-                     ++actionIndex) {
+                for (size_t actionIndex = 0; actionIndex < numberOfActions; ++actionIndex) {
                     if (actionIsApplicable(actionStates[actionIndex], state)) {
                         // This action is applicable
                         PDState nxt(state.remainingSteps() - 1);
@@ -351,7 +347,7 @@ protected:
                     }
                 }
             } else {
-                for (unsigned int actionIndex = 0; actionIndex < numberOfActions; ++actionIndex) {
+                for (size_t actionIndex = 0; actionIndex < numberOfActions; ++actionIndex) {
                     if (actionIsApplicable(actionStates[actionIndex], state)) {
                         res[actionIndex] = actionIndex;
                     } else {
@@ -371,7 +367,7 @@ protected:
     std::vector<int> getIndicesOfApplicableActions(State const& state) const {
         std::vector<int> applicableActions = getApplicableActions(state);
         std::vector<int> result;
-        for (unsigned int actionIndex = 0; actionIndex < applicableActions.size();
+        for (size_t actionIndex = 0; actionIndex < applicableActions.size();
              ++actionIndex) {
             if (applicableActions[actionIndex] == actionIndex) {
                 result.push_back(actionIndex);
@@ -388,7 +384,7 @@ protected:
     void calcKleeneSuccessor(KleeneState const& current,
                              int const& actionIndex,
                              KleeneState& next) const {
-        for (unsigned int i = 0; i < KleeneState::stateSize; ++i) {
+        for (size_t i = 0; i < KleeneState::stateSize; ++i) {
             allCPFs[i]->evaluateToKleene(next[i], current, 
                                          actionStates[actionIndex]);
         }
@@ -447,24 +443,26 @@ protected:
 
     // Apply action 'actionIndex' in the determinization to 'current', get state
     // 'next' and yield reward 'reward'
-    void calcStateTransition(State const& current, int const& actionIndex,
-            State& next,
-            double& reward) const {
+    void calcStateTransition(State const& current,
+                             int const& actionIndex,
+                             State& next,
+                             double& reward) const {
         calcSuccessorState(current, actionIndex, next);
         calcReward(current, actionIndex, reward);
     }
 
     // Apply action 'actionIndex' in the determinization to 'current', resulting
     // in 'next'.
-    void calcSuccessorState(State const& current, int const& actionIndex,
+    void calcSuccessorState(State const& current,
+                            int const& actionIndex,
             State& next) const {
-        for (unsigned int index = 0; index < State::numberOfDeterministicStateFluents; ++index) {
+        for (size_t index = 0; index < State::numberOfDeterministicStateFluents; ++index) {
             deterministicCPFs[index]->evaluate(next.deterministicStateFluent(index),
                                                current,
                                                actionStates[actionIndex]);
         }
 
-        for (unsigned int index = 0; index < State::numberOfProbabilisticStateFluents; ++index) {
+        for (size_t index = 0; index < State::numberOfProbabilisticStateFluents; ++index) {
             determinizedCPFs[index]->evaluate(next.probabilisticStateFluent(index),
                                               current,
                                               actionStates[actionIndex]);
@@ -489,14 +487,14 @@ protected:
         ActionHashMap::iterator it = applicableActionsCache.find(state);
         if (it != applicableActionsCache.end()) {
             assert(it->second.size() == res.size());
-            for (unsigned int i = 0; i < res.size(); ++i) {
+            for (size_t i = 0; i < res.size(); ++i) {
                 res[i] = it->second[i];
             }
         } else {
             if (hasUnreasonableActions) {
                 std::map<State, int, State::CompareIgnoringRemainingSteps> childStates;
 
-                for (unsigned int actionIndex = 0; actionIndex < numberOfActions; ++actionIndex) {
+                for (size_t actionIndex = 0; actionIndex < numberOfActions; ++actionIndex) {
                     if (actionIsApplicable(actionStates[actionIndex], state)) {
                         // This action is applicable
                         State nxt;
@@ -517,7 +515,7 @@ protected:
                     }
                 }
             } else {
-                for (unsigned int actionIndex = 0; actionIndex < numberOfActions; ++actionIndex) {
+                for (size_t actionIndex = 0; actionIndex < numberOfActions; ++actionIndex) {
                     if (actionIsApplicable(actionStates[actionIndex], state)) {
                         res[actionIndex] = actionIndex;
                     } else {
@@ -536,7 +534,7 @@ protected:
     std::vector<int> getIndicesOfApplicableActions(State const& state) const {
         std::vector<int> applicableActions = getApplicableActions(state);
         std::vector<int> result;
-        for (unsigned int actionIndex = 0; actionIndex < applicableActions.size();
+        for (size_t actionIndex = 0; actionIndex < applicableActions.size();
              ++actionIndex) {
             if (applicableActions[actionIndex] == actionIndex) {
                 result.push_back(actionIndex);
